requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
python-docx>=1.1.0
docx2txt>=0.8
python-dotenv>=1.0.0
streamlit>=1.40.0
